requirements.txt
grpcio==1.9.1
grpcio-tools==1.9.1
protobuf==3.5.1
pyinotify
